requirements.txt
datasets
datasketch
faiss_cpu
geopandas
jarowinkler
joblib
numba
numpy
pandas
pynndescent
rapidfuzz
scikit_learn
scipy
setuptools
tokenizers
torch
tqdm
transformers
